undo renaming utils_tags
